Added option to disable 2FA
This mod introduces DISABLE_2FA parameter in [security] section
of app.ini (by default set to false). If set to true it disables access
to 2FA feature in user preferences (not required in some environments
i.e. when reverse proxy auth is used). Authentication code using 2FA and
any existing 2FA configuration are left untouched.

This mod hides also security tab in user preferences when openid is also
disabled; for this reason this mod is not separate PR but exiting PR
enhancement.

Author-Change-Id: IB#1105071